Add adition functionality for post author

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,7 @@ use App\Models\Post;
 use App\Models\Topic;
 //use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage; //Facade for storage operations
 
 class PostController extends Controller
 {
@@ -16,14 +17,19 @@ class PostController extends Controller
         return view('post.create', compact('topics'));
     }
 
-/*     public function show(Post $post)
+    public function show(Post $post)
     {
-        return view()
+        $topics = Topic::all('title', 'id');
+        return view('post.show', compact('post', 'topics'));
     }
- */
-    public function store(StorePostRequest $request)
+
+    public function update(StorePostRequest $request, Post $post)
     {
-        // Определяем тему
+        if ($post->user_id !== Auth::id()) {
+            abort(403);
+        }
+
+        // Define topic
         if ($request->topic_mode === 'new') {
 
             $topic = Topic::create([
@@ -35,7 +41,44 @@ class PostController extends Controller
             $topic = Topic::findOrFail($request->existing_topic_id);
         }
 
-        // Обработка изображения
+        $imagePath = $post->picture; //Default - keep last picture without changing
+
+        //If $request has new picture
+        //delete old one and save new
+        if ($request->hasFile('userPicture')) {
+
+            // delete old image
+            if ($post->picture) {
+                Storage::disk('public')->delete($post->picture);
+            }
+
+            // save new one
+            $imagePath = $request->file('userPicture')->store('post-images', 'public');
+        }
+
+        $post->update([
+            'content' => $request->content,
+            'topic_id' => $topic->id,
+            'picture' => $imagePath,
+        ]);
+
+        return redirect()->route('home.index')->with('success', 'Your post has been updated');
+    }
+
+    public function store(StorePostRequest $request)
+    {
+        // Define topic
+        if ($request->topic_mode === 'new') {
+
+            $topic = Topic::create([
+                'title' => $request->new_topic_title,
+                'user_id' => Auth::id(),
+            ]);
+        } else {
+            $topic = Topic::findOrFail($request->existing_topic_id);
+        }
+
+        // Post processing
         $imagePath = null;
 
         if ($request->hasFile('userPicture')) {
@@ -43,7 +86,7 @@ class PostController extends Controller
                 ->store('post-images', 'public');
         }
 
-        // Создание поста
+        // Creation post
         Post::create([
             'content' => $request->content,
             'user_id' => Auth::id(),
@@ -51,13 +94,17 @@ class PostController extends Controller
             'picture' => $imagePath,
         ]);
 
-        return redirect()
-            ->route('home.index')
+        return redirect()->route('home.index')
             ->with('success', 'Post created successfully!');
     }
 
+
     public function destroy(Post $post)
     {
+        if ($post->user_id !== Auth::id()) {
+            abort(403);
+        }
+        
         $post->delete();
 
         return redirect()->route('home.index');
